Add Command class with execute method
This is a contract for all the commands to adhere to, as they all have
execute called on them inside Bitmapper. I don't know if this is a common
thing to do in Ruby; it still does not enforce the contract in the same
way an interface would in Java. I can still include an object in the list
of commands that does not extend the Command class. However I think this
is an improvement as it makes the design clearer.

Also update README with description of how it works.